+++ README.md
@@ -68,22 +68,24 @@ Usage:
   trojan [command]
 
 Available Commands:
-  add         添加用户
-  clean       清空指定用户流量
-  completion  自动命令补全(支持bash和zsh)
-  del         删除用户
-  help        Help about any command
-  info        用户信息列表
-  log         查看trojan日志
-  restart     重启trojan
-  start       启动trojan
-  status      查看trojan状态
-  stop        停止trojan
-  tls         证书安装
-  update      更新trojan
-  updateWeb   更新trojan管理程序
-  version     显示版本号
-  web         以web方式启动
+  add           添加用户
+  clean         清空指定用户流量
+  completion    自动命令补全(支持bash和zsh)
+  del           删除用户
+  help          Help about any command
+  info          用户信息列表
+  log           查看trojan日志
+  restart       重启trojan
+  start         启动trojan
+  status        查看trojan状态
+  stop          停止trojan
+  tls           证书安装
+  update        更新trojan
+  updateWeb     更新trojan管理程序
+  version       显示版本号
+  import [path] 导入sql文件
+  export [path] 导出sql文件
+  web           以web方式启动
 
 Flags:
   -h, --help   help for trojan
