+++ trojan/trojan.go
@@ -152,23 +152,3 @@ func LogChan(param string, closeChan chan byte) (chan string, error) {
 	}()
 	return ch, nil
 }
-
-// SetDomain 设置显示的域名
-func SetDomain(domain string) {
-	if domain == "" {
-		domain = util.Input("请输入要显示的域名地址: ", "")
-	}
-	if domain == "" {
-		fmt.Println("撤销更改!")
-	} else {
-		core.WriteDomain(domain)
-		Restart()
-		fmt.Println("修改domain成功!")
-	}
-}
-
-// GetDomainAndPort 获取域名和端口
-func GetDomainAndPort() (string, int) {
-	config := core.Load("")
-	return config.SSl.Sni, config.LocalPort
-}
